+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Version 10.0.5.2
 
-* Added options to hide slot artwork for Bank, Guild Bank, Reagent Bank, and Void Storage
+* Added options to hide slot artwork for Bank, Guild Bank, Reagent Bank, Void Storage, and Mail
 * Added missing Open Mail Money Button
+* Fixed border colors for Mail Inbox buttons
+  * Note that the highlight that appears when a Mail item is open isn't skinned correctly
+  * This should be fixed in an upcoming release of Masque (10.0.6)
 * Significantly reorganized the code
   * Split functionality into separate files
   * Improved generalization and removed workarounds for Mail Items